@@ -5,11 +5,9 @@
   "author": "terentev m",
   "private": true,
   "scripts": {
-    "dev": "webpack-dev-server --inline --progress --config build/webpack.dev.conf.js",
-    "start": "npm run dev",
-    "build": "node build/build.js",
-    "prettier-js": "prettier --write src/**/*.js",
-    "prettier-vue": "prettier --write src/**/*.vue"
+    "dev": "cross-env NODE_ENV=development webpack-dev-server --progress",
+    "staging": "cross-env NODE_ENV=staging webpack",
+    "build": "cross-env NODE_ENV=production webpack"
   },
   "dependencies": {
     "@babel/runtime": "^7.5.5",
@@ -18,11 +16,14 @@
     "babel-polyfill": "^6.26.0",
     "bootstrap": "^4.3.1",
     "chart.js": "^2.8.0",
+    "material-icons": "^0.3.1",
     "moment": "^2.24.0",
+    "v-calendar": "^1.0.0-beta.22",
     "vee-validate": "^2.2.15",
     "vue": "^2.5.22",
     "vue-chartjs": "^3.4.2",
     "vue-router": "^3.0.2",
+    "vue-select": "^3.2.0",
     "vuetify": "^1.5.16",
     "vuetify-loader": "^1.3.0",
     "vuex": "^3.1.0"
@@ -52,7 +53,9 @@
     "extract-text-webpack-plugin": "^3.0.0",
     "file-loader": "^1.1.11",
     "friendly-errors-webpack-plugin": "^1.7.0",
-    "html-webpack-plugin": "^2.30.1",
+    "html-critical-webpack-plugin": "^2.1.0",
+    "html-webpack-plugin": "~3.2",
+    "mini-css-extract-plugin": "^0.8.0",
     "moment-locales-webpack-plugin": "^1.1.0",
     "node-notifier": "^5.3.0",
     "node-sass": "^4.11.0",
@@ -73,13 +76,17 @@
     "stylus-loader": "^3.0.2",
     "uglifyjs-webpack-plugin": "^1.3.0",
     "url-loader": "^0.5.8",
-    "vue-loader": "^13.7.3",
+    "vue-loader": "~15.6",
     "vue-style-loader": "^3.1.2",
     "vue-template-compiler": "^2.5.22",
-    "webpack": "^3.12.0",
-    "webpack-bundle-analyzer": "^2.13.1",
-    "webpack-dev-server": "^2.11.3",
-    "webpack-merge": "^4.2.1"
+    "compression-webpack-plugin": "~2.0",
+    "cross-env": "~5.2",
+    "webpack": "~4.29",
+    "webpack-bundle-analyzer": "~3.3",
+    "webpack-cli": "~3.2",
+    "webpack-dev-server": "~3.1",
+    "webpack-hot-middleware": "~2.24",
+    "webpack-merge": "~4.2"
   },
   "engines": {
     "node": ">= 6.0.0",
