@@ -1,32 +1,22 @@
 <template>
-    <v-hover>
-        <v-card slot-scope="{ hover }" class="mx-1">
-            <v-card-title
-                    class="dash-card-header-title"
 
-            >
+        <div class="card mt-1">
+            <div class="card-header bg-transparent dash-card-header-title">
                 Balance
-            </v-card-title>
-            <v-divider light></v-divider>
+            </div>
 
-            <v-layout>
-                <v-flex xs7>
-                    <v-card-title >
-                        <div>
-                            <div class="headline">{{ balanceRounded }}</div>
-                            <div class="grey--text">{{ balanceCurrency.short_name }}</div>
-                        </div>
-                    </v-card-title>
-                </v-flex>
+            <div class="card-body d-flex flex-row justify-content-between align-items-center">
+                <div>
+                    <div class="headline">{{ balanceRounded }}</div>
+                    <div class="grey--text">{{ balanceCurrency.short_name }}</div>
+                </div>
 
-                <v-flex xs5 d-flex justify-space-around>
-                    <img src="../../assets/balance.svg" height="64px" width="64px"/>
-                </v-flex>
-            </v-layout>
+                <img src="../../assets/balance.svg" height="64px" width="64px" alt="...">
 
+            </div>
 
-        </v-card>
-    </v-hover>
+        </div>
+
 </template>
 
 <script>
