@@ -1,26 +1,22 @@
 <template>
-    <v-card class="mx-1">
-        <v-card-title class="dash-card-header-title">
+
+    <div class="card mt-1">
+        <div class="card-header bg-transparent dash-card-header-title">
             Incomes
-    </v-card-title>
+        </div>
 
-    <v-divider light></v-divider>
+        <div class="card-body d-flex flex-row justify-content-between align-items-center">
+            <div>
+                <div class="headline">{{ incomesRounded }}</div>
+                <div class="grey--text">{{ totalIncomesCurrency.short_name }}</div>
+            </div>
 
-    <v-layout>
-      <v-flex xs7>
-          <v-card-title>
-          <div>
-              <div class="headline">{{ incomesRounded }}</div>
-              <div class="grey--text">{{ totalIncomesCurrency.short_name }}</div>
-          </div>
-        </v-card-title>
-      </v-flex>
+            <img src="../../assets/graph_income.svg" height="64px" width="64px">
 
-      <v-flex xs5 d-flex justify-space-around>
-        <img src="../../assets/graph_income.svg" height="64px" width="64px" />
-      </v-flex>
-    </v-layout>
-  </v-card>
+        </div>
+
+    </div>
+
 </template>
 
 <script>
@@ -29,10 +25,10 @@
     export default {
         props: ['totalIncomes', 'totalIncomesCurrency'],
 
-  computed: {
-    incomesRounded: function() {
-        return Mynum.round2(this.totalIncomes);
-    },
-  },
-};
+        computed: {
+            incomesRounded: function () {
+                return Mynum.round2(this.totalIncomes);
+            },
+        },
+    };
 </script>
