@@ -1,103 +1,134 @@
 <template>
-    <v-container>
-        <v-layout row>
-            <v-flex xs12>
-                <div class="table-wrapper">
-                    <v-layout row mx-2>
-                        <v-flex xs12 sm12 md6 lg6 xl6>
-                            <v-card flat>
-                                <h4 class="card-subtitle mt-3 mb-3 text-muted">General</h4>
-                                <v-form>
-                                    <v-select
+    <div class="container">
 
-                                            dense
-                                            :value="currency"
-                                            :items="currencies"
-                                            label="System currency"
-                                            readonly
-                                            item-text="name"
-                                            item-value="id"
-                                            return-object
-                                    ></v-select>
-                                    <v-autocomplete
-                                            v-model="wallet"
-                                            :value="wallet"
-                                            :items="wallets"
-                                            label="Main wallet"
-                                            clearable
-                                            item-text="name"
-                                            item-value="id"
-                                            return-object
-                                    ></v-autocomplete>
-                                    <v-autocomplete
-                                            v-model="currencyReports"
-                                            :value="currencyReports"
-                                            :items="currencies"
-                                            label="Currency for reports"
-                                            clearable
-                                            item-text="name"
-                                            item-value="id"
-                                            return-object
-                                    ></v-autocomplete>
-                                </v-form>
-                                <v-card-actions>
-                                    <v-progress-linear v-show="this.processing"
-                                                       :indeterminate="true">
-                                        <p>process</p>
-                                    </v-progress-linear>
-                                    <v-spacer></v-spacer>
+        <div class="row">
 
-                                    <v-btn
-                                            type="button"
-                                            :disabled="this.processing === true"
-                                            v-if="formModified === true"
-                                            @click="storeChanges"
-                                            color="info">Save changes
-                                    </v-btn>
-                                </v-card-actions>
-                            </v-card>
-                        </v-flex>
+            <div class="col-xs-12 col-sm-12 col-md-5 ">
+                <div class="row d-block">
 
-                        <v-flex xs12 sm12 md6 lg6 xl6>
-                            <v-card flat mx-2>
-                                <h4 class="card-subtitle mt-3 mb-3 text-muted">Dashboard
-                                    settings</h4>
-                                <v-select
-                                        :items="periods"
-                                        v-model="periodicity"
-                                        item-text="name"
-                                        item-value="period"
-                                        return-object
-                                        label="Periodicity"
-                                        prepend-icon="date_range"
-                                ></v-select>
-                            </v-card>
-                        </v-flex>
+                    <div class="form-group">
+                        <label for="settings_currency_el" class="tm-lable">Currency (read only):</label>
 
-                    </v-layout>
+                        <tm-select
+                                id="settings_currency_el"
+                                v-model="currency"
+                                :options="currencies"
+                                :title="'name'"
+                                :clearable="true"
+                                :select-btn="true"
+                                :placeholder="'Select main wallet'"
+                                :read-only="true"
+                        ></tm-select>
+
+                    </div>
                 </div>
-            </v-flex>
 
-        </v-layout>
+                <div class="row d-block">
+                    <div class="form-group">
+                        <label for="settings_main_wallet_el" class="tm-lable">Default wallet<span v-show="isDirty('wallet')"> *</span>:</label>
 
-        <v-snackbar v-model="this.showAlert" vertical :color="this.alertType">
-            {{ this.message }}
-            <v-btn flat @click="showAlert = false">Close</v-btn>
+                        <tm-select
+                                id="settings_main_wallet_el"
+                                v-model="wallet"
+                                :options="wallets"
+                                :title="'name'"
+                                :clearable="true"
+                                :select-btn="true"
+                                :placeholder="'Select main wallet'"
+                                @open="openWalletSelectForm"
+                        ></tm-select>
+
+                    </div>
+                </div>
+
+                <div class="row d-block ">
+                    <div class="form-group">
+                        <label for="settings_rep_currency_el" class="tm-lable">Currency of dashboard<span v-show="isDirty('currencyReports')"> *</span>:</label>
+                        <tm-select
+                                v-model="currencyReports"
+                                id="settings_rep_currency_el"
+                                :options = "currencies"
+                                :title="'name'"
+                                :clearable="true"
+                                :select-btn="true"
+                                :placeholder="'Select currency for reports'"
+                        ></tm-select>
+                    </div>
+                </div>
+
+                <div class="row d-flex px-0 px-sm-3">
+                    <button
+                            v-if="formModified"
+                            type="button"
+                            :disabled="this.processing"
+                            class="btn btn-success"
+                            @click="storeChanges"
+                    >Store changes</button>
+                </div>
+                <v-progress-linear v-show="this.processing" :indeterminate="true">
+                </v-progress-linear>
+            </div>
+
+            <div class="col-xs-0 col-sm-0 col-md-2 "></div>
+
+            <tm-wallets-select-form
+                    :items="this.wallets"
+                    :showWalletSelection="this.showWalletSelection"
+                    @select-wallets-close="completeWalletSelectionHandler"
+            ></tm-wallets-select-form>
+            <div class="col-xs-12 col-sm-12 col-md-5">
+                <div class="row d-block">
+
+                    <div class="form-group">
+                        <label for="settings_dash_period_el" class="tm-lable">Periodicity:</label>
+                        <tm-select
+                                v-model="periodicity"
+                                id="settings_dash_period_el"
+                                :options = "periods"
+                                :title="'period'"
+                                :clearable="true"
+                                :select-btn="false"
+                                :placeholder="'Set periodicity for the app dashboard'"
+                        ></tm-select>
+                    </div>
+                </div>
+            </div>
+        </div>
+
+        <v-snackbar
+            v-model="showAlert"
+            :color="snackBarColor()"
+
+        >
+            {{ message }}
+            <v-btn color="white" flat @click="showAlert = false">Close</v-btn>
         </v-snackbar>
-    </v-container>
+
+    </div>
+
+
 </template>
 
 <script>
     import {mapState} from 'vuex';
-
+    import {VProgressLinear} from 'vuetify/lib/components/VProgressLinear';
+    import {VSnackbar} from 'vuetify/lib/components/VSnackbar';
     export default {
         data: () => ({
             formModified: false,
+            showWalletSelection: false,
             processing: false,
             showAlert: false,
             alertType: 'success',
             message: '',
+
+            dirtyElements: []
         }),
+
+        components: {
+            VProgressLinear,
+            VSnackbar
+        },
 
         computed: {
             ...mapState({
@@ -113,7 +144,13 @@
                 }
                 ,
                 set(currency) {
+
                     this.formModified = true;
+
+                    if (!this.dirtyElements.includes('currencyReports')) {
+                        this.dirtyElements.push('currencyReports');
+                    }
+
                     this.$store.commit('setCurrencyReports', currency);
                 }
             },
@@ -123,6 +160,9 @@
                 },
                 set(wallet) {
                     this.formModified = true;
+                    if (!this.dirtyElements.includes('wallet')) {
+                        this.dirtyElements.push('wallet');
+                    }
                     this.$store.commit('setWallet', wallet);
                 }
             },
@@ -149,19 +189,58 @@
         },
 
         methods: {
+            openWalletSelectForm() {
+                this.showWalletSelection = true;
+            },
+
+            completeWalletSelectionHandler(wallet) {
+                if (wallet instanceof Object) {
+                    this.formModified = true;
+                    this.$store.commit('setWallet', wallet);
+
+                }
+
+                this.showWalletSelection = false;
+            },
+
+            isDirty(elemName) {
+                return this.dirtyElements.includes(elemName);
+            },
+
+            snackBarColor() {
+                return this.alertType == 'success' ? 'success' : 'error';
+            },
+
             storeChanges() {
                 this.processing = true;
                 this.showAlert = false;
                 this.$store.dispatch('storeSettings').then(() => {
                     this.processing = false;
+                    this.formModified = false;
+                    this.dirtyElements = [];
                     this.message = 'Stored successfully';
                     this.alertType = 'success';
                     this.showAlert = true;
 
                 }).catch(err => {
+
+                    let status = 'n/a';
+                    let errorText = 'Error';
+
+                    if (err instanceof Object && err.hasOwnProperty('status')) {
+                        status = err.status;
+                    }
+
+                    if (err instanceof Object && err.hasOwnProperty('data')) {
+                        if (err.data.hasOwnProperty('error')) {
+                            errorText = err.data.error;
+                        }
+
+                    }
+
                     this.processing = false;
 
-                    this.message = 'Error';
+                    this.message = `(${status}) ${errorText}`;
                     this.alertType = 'error';
                     this.showAlert = true;
                 });
@@ -173,3 +252,11 @@
 
 
 </script>
+<style>
+    .input-select-wrapper {
+        display: flex;
+        flex-basis: 400px;
+        width: 100%;
+        flex-grow: 1;
+    }
+</style>
