@@ -1,85 +1,66 @@
 import 'babel-polyfill';
 import Vue from 'vue';
 import App from './App';
-import router from './router';
 
-import Vuetify, {
-    VAlert,
-    VApp,
-    VAutocomplete,
-    VAvatar,
-    VBottomSheet,
-    VBtn,
-    VCard,
-    VCardActions,
-    VCardText,
-    VCardTitle,
-    VCheckbox,
-    VChip,
-    VContainer,
-    VContent,
-    VDatePicker,
-    VDialog,
-    VDivider,
-    VFlex,
-    VForm,
-    VHover,
-    VIcon,
-    VInput,
-    VItem,
-    VItemGroup,
-    VLabel,
-    VLayout,
-    VList,
-    VListGroup,
-    VListTile,
-    VListTileAction,
-    VListTileActionText,
-    VListTileAvatar,
-    VListTileContent,
-    VListTileSubTitle,
-    VListTileTitle,
-    VMenu,
-    VMessages,
-    VNavigationDrawer,
-    VProgressCircular,
-    VProgressLinear,
-    VSelect,
-    VSnackbar,
-    VSpacer,
-    VSubheader,
-    VSystemBar,
-    VTextarea,
-    VTextField,
-    VToolbar,
-    VToolbarItems,
-    VToolbarSideIcon,
-    VToolbarTitle,
-} from 'vuetify/lib';
+import router from './router';
+import store from './store';
+
+// Import Vuetify components ++
+import Vuetify from 'vuetify/lib/components/Vuetify';
+
+import VAvatar from 'vuetify/lib/components/VAvatar';
+import VBtn from 'vuetify/lib/components/VBtn';
+import VCard from 'vuetify/lib/components/VCard';
+import VCardActions from 'vuetify/lib/components/VCard';
+import VCardText from 'vuetify/lib/components/VCard';
+import VCardTitle from 'vuetify/lib/components/VCard';
+import VCheckbox from 'vuetify/lib/components/VCheckbox';
+import VContainer from 'vuetify/lib/components/VGrid';
+import VDialog from 'vuetify/lib/components/VDialog';
+import VDivider from 'vuetify/lib/components/VDivider';
+import VFlex from 'vuetify/lib/components/VGrid';
+import VIcon from 'vuetify/lib/components/VIcon';
+import VItem from 'vuetify/lib/components/VItemGroup';
+import VItemGroup from 'vuetify/lib/components/VItemGroup';
+import VLabel from 'vuetify/lib/components/VLabel';
+import VLayout from 'vuetify/lib/components/VGrid';
+import VList from 'vuetify/lib/components/VList';
+import VListGroup from 'vuetify/lib/components/VList';
+import VListTile from 'vuetify/lib/components/VList';
+import VListTileAction from 'vuetify/lib/components/VList';
+import VListTileActionText from 'vuetify/lib/components/VList';
+import VListTileAvatar from 'vuetify/lib/components/VList';
+import VListTileContent from 'vuetify/lib/components/VList';
+import VListTileSubTitle from 'vuetify/lib/components/VList';
+import VListTileTitle from 'vuetify/lib/components/VList';
+import VMessages from 'vuetify/lib/components/VMessages';
+import VProgressLinear from 'vuetify/lib/components/VProgressLinear'; //Global
+import VSnackbar from 'vuetify/lib/components/VSnackbar';
+import VSpacer from 'vuetify/lib/components/VGrid';
+import VSubheader from 'vuetify/lib/components/VSubheader';
+import VToolbar from 'vuetify/lib/components/VToolbar';
+import VToolbarItems  from 'vuetify/lib/components/VToolbar';
+import VToolbarSideIcon from 'vuetify/lib/components/VToolbar';
+import VToolbarTitle from 'vuetify/lib/components/VToolbar';
+
+// Import Vuetify components --
+
+//Import own components ++
+import TM_WalletsSelectForm from './components/SelectsForms/WalletsSelectForm.vue';
+import TMSelect from './components/TMComponents/TMSelect/TMSelect';
+//Import own components --
 
 import {install as VeeValidate, Validator} from 'vee-validate/dist/vee-validate.minimal.esm.js';
 import {email, max, min, required} from 'vee-validate/dist/rules.esm.js';
 import veeEn from 'vee-validate/dist/locale/en';
-//my custom list
-import RoundedList from './components/Reusable/ListOfAmount.vue';
-import TMDataTable from './components/Reusable/TMDataTable/TMDataTable';
 
-import BalanceTotal from './components/IndexDashboard/BalanceTotal.vue';
-import BalanceCard from './components/IndexDashboard/BalanceCard.vue'
-import ExpenseTotal from './components/IndexDashboard/ExpenseTotal.vue';
-import IncomeTotal from './components/IndexDashboard/IncomeTotal.vue';
-import ExpensesList from './components/IndexDashboard/ExpensesList.vue';
-import IncomesList from './components/IndexDashboard/IncomesList.vue';
-import BalanceByWallets from './components/IndexDashboard/BalanceByWallets.vue';
-import TM_FormActionPanel from './components/tmFormElements/TM_FormActionPanel.vue';
-import TM_WalletsSelectForm from './components/SelectsForms/WalletsSelectForm.vue';
 
-import store from './store';
-
+//Import styles ++
+import 'bootstrap/dist/css/bootstrap.min.css';
 import 'vuetify/src/stylus/app.styl'
 import "../src/assets/css/sprite.css";
 import './style/pf.css';
-import 'bootstrap/dist/css/bootstrap.min.css';
+//Import styles --
 
 import globalFunctions from './mixins/GlobalFunctions';
 
@@ -101,29 +82,19 @@ Vue.use(Vuetify, {
         appColor: mainColor,
     },
     components: {
-        VApp, // required
         VAvatar,
         VContainer,
         VFlex,
         VLayout,
-        VContent,
-        VAlert,
-        VAutocomplete,
         VBtn,
-        VBottomSheet,
         VCard,
         VCardTitle,
         VCardText,
         VCardActions,
         VCheckbox,
-        VChip,
-        VDatePicker,
         VDialog,
         VDivider,
-        VForm,
-        VHover,
         VIcon,
-        VInput,
         VItem,
         VItemGroup,
         VLabel,
@@ -136,40 +107,24 @@ Vue.use(Vuetify, {
         VListTileContent,
         VListTileSubTitle,
         VListTileTitle,
-        VMenu,
         VMessages,
-        VNavigationDrawer,
-        VProgressCircular,
         VProgressLinear,
-        VSelect,
         VSnackbar,
         VSpacer,
         VSubheader,
-        VSystemBar,
-        VTextarea,
-        VTextField,
         VToolbar,
         VToolbarSideIcon,
         VToolbarTitle,
         VToolbarItems,
-
     }
 });
 
 
-Vue.component('tm-table', TMDataTable);
-Vue.component('tm-list-rounded', RoundedList);
-Vue.component('tm-index-balance-card', BalanceCard);
-Vue.component('tm-index-balance-total', BalanceTotal);
-Vue.component('tm-index-expense-total', ExpenseTotal);
-Vue.component('tm-index-income-total', IncomeTotal);
-Vue.component('tm-index-expenses-list', ExpensesList);
-Vue.component('tm-index-incomes-list', IncomesList);
-Vue.component('tm-index-balance-list', BalanceByWallets);
-Vue.component('tm-form-action-panel', TM_FormActionPanel);
 Vue.component('tm-wallets-select-form', TM_WalletsSelectForm);
+Vue.component('tm-select', TMSelect);
 
 Vue.config.productionTip = false;
+
 
 
 Vue.mixin(globalFunctions);
@@ -195,14 +150,10 @@ new Vue({
     el: '#app',
     router: router,
     store,
-
     data: {
         visible: true,
-        message: 'Hello Vue!',
         showMenu: true,
     },
-
     template: '<App/>',
-
     components: {App}
 });
